Move RequireColumnsVisitor to planners
