+++ query/src/optimizers/mod.rs
@@ -29,4 +29,3 @@ pub use optimizer_scatters::ScattersOptimizer;
 pub use optimizer_statistics_exact::StatisticsExactOptimizer;
 pub use optimizer_top_n_push_down::TopNPushDownOptimizer;
 pub use utils::MonotonicityCheckVisitor;
-pub use utils::RequireColumnsVisitor;
